Implement the data archival procedure
This commit introduces the archival procedure, which is one of the two
core functionalities of this project (the other being the erasure).

The procedure will seek every known GShade file and back it up to a ZIP
archive with the following top-level directories:

  Name   Description
 ------ --------------------------------------------------------
  main   The shaders and presets -- this is what matters most!
  hook   DLL files used by GShade (ReShade) to inject onto XIV
  core   Installation files scattered across the filesystem
  misc   Shortcuts and other goodies which depend on GShade

By default, the archive name will be timestamped. The path and name can
be overriden using the --archive parameter.

Signed-off-by: Emilian Roman <git@miris.design>