Revise order of path detection (prioritise current directory)
This commit revises the order of how the GShade & XIV path detection
procedures are executed.

Previously, the GShade registry key would be checked first, as it would
usually offer the most reliable value. Now, it is checked last, and the
working directory is checked first. This is done for two reasons:

1.  Many people running GShade.Nuke would've likely tried uninstalling
    GShade using its official uninstaller, thereby deleting the key in
    the registry.

2.  If GShade.Nuke is present in the XIV 'game' directory, it would've
    been a deliberate attempt at specfying the path. The same applies
    when the program is placed in the GShade installation path.

Signed-off-by: Emilian Roman <git@miris.design>